feat: add `groupId` and `sync` options
