+++ .yarn/versions/1469515d.yml
@@ -1,0 +1,2 @@
+releases:
+  docusaurus-remark-plugin-tab-blocks: minor
