+++ src/__tests__/index.test.js
@@ -1,0 +1,5 @@
+import tabBlocksPlugin from "../../";
+
+test("tabBlocksPlugin", () => {
+  expect(tabBlocksPlugin).toBeDefined();
+});
